Qt: Fix implicit conversion
Fix

ui/qt/widgets/syntax_line_edit.cpp:160:51: warning: implicit conversion loses integer precision: 'unsigned long' to 'int' [-Wshorten-64-to-32]
            .arg(QString('~').repeated(loc_length - 1));
                              ~~~~~~~~ ~~~~~~~~~~~^~~
ui/qt/widgets/syntax_line_edit.cpp:159:40: warning: implicit conversion loses integer precision: 'qsizetype' (aka 'long long') to 'int' [-Wshorten-64-to-32]
            .arg(QString(' ').repeated(loc_start))
                              ~~~~~~~~ ^~~~~~~~~

when building with Qt 5.

+++ ui/qt/widgets/syntax_line_edit.cpp
@@ -156,8 +156,8 @@ QString SyntaxLineEdit::createSyntaxErrorMessageFull(
         msg = QString("<p>%1<pre>  %2\n  %3^%4</pre></p>")
             .arg(msg)
             .arg(filter)
-            .arg(QString(' ').repeated(loc_start))
-            .arg(QString('~').repeated(loc_length - 1));
+            .arg(QString(' ').repeated(static_cast<int>(loc_start)))
+            .arg(QString('~').repeated(static_cast<int>(loc_length) - 1));
     }
     return msg;
 }
